CumulativeProfile: ensure correct type

+++ misura/canon/reference/profile.py
@@ -42,7 +42,7 @@ class Profile(VariableLength):
             return None
         if len(prf[0]) != 2:
             return None
-        (w, h), x, y = prf
+        (w, h), x, y = prf.astype(np.uint16)
         # Cast double time into four 16-bit integers
         ta = np.array(binary_cast([t], 'd', 'HHHH'))
         # Create the concatenated arrays
@@ -97,11 +97,11 @@ def couple(d, scale=9):
         d = np.concatenate((d, [4]))
     #  Sum even with odd by applying a scaling
     d1 = d[0::2] + d[1::2]*scale
-    return d1.astype('uint8')
+    return d1
 
 def decouple(v, scale=9):
     # Conversion to bigger SIGNED int
-    v=v.astype('int16')
+    v=v.astype(np.uint16)
     # Unpack bits from bytes
     v_even = v % scale
     v_odd = v // scale
@@ -167,7 +167,7 @@ class CumulativeProfile(Profile):
         (w, h), x, y = prf
         # Cast double time,  16-bits integer dimensions and starting coords into 16 8-bits unsigned integers
         coords = np.array(binary_cast([t, w, h, x[0], y[0]], 'dHHHH', '<16B'))
-        cumul = accumulate_coords(x,y)
+        cumul = accumulate_coords(x.astype(np.uint8),y.astype(np.uint8))
         out = np.concatenate((coords, cumul))
         return out
